Replaced craft action description results.

+++ app/classes/action.rb
@@ -373,7 +373,7 @@ Action.new("craft_cutter",
   { :name=>"Craft Cutter",
     :description=>"Craft a simple cutting tool to aid in harvesting.",
     :result => lambda { |character, character_action|
-      if(character.possesses?("tomatunk") && character.posseses?("dolait"))
+      if(character.possesses?("tomatunk") && character.possesses?("dolait"))
         character.character_possessions.where(:possession_id=>"tomatunk").first.destroy!
         if(Random.rand(2)==0)
           CharacterPossession.new(:character_id => character.id, :possession_id => "cutter").save!
@@ -397,7 +397,7 @@ Action.new("craft_shaper_a",
   { :name=>"Craft Oblong Shaper",
     :description=>"Craft a simple oblong shaping tool to aid in crafting.",
     :result => lambda { |character, character_action|
-      if(character.possesses?("tomatunk") && character.posseses?("dolait"))
+      if(character.possesses?("tomatunk") && character.possesses?("dolait"))
         character.character_possessions.where(:possession_id=>"dolait").first.destroy!
         character.character_possessions.where(:possession_id=>"tomatunk").first.destroy!
         if(Random.rand(2)==0)
@@ -422,17 +422,17 @@ Action.new("craft_shaper_b",
   { :name=>"Craft Angled Shaper",
     :description=>"Craft a simple angled shaping tool to aid in crafting.",
     :result => lambda { |character, character_action|
-      if(character.possesses?("tomatunk") && character.posseses?("dolait"))
+      if(character.possesses?("tomatunk") && character.possesses?("dolait"))
         character.character_possessions.where(:possession_id=>"dolait").first.destroy!
         character.character_possessions.where(:possession_id=>"tomatunk").first.destroy!
         CharacterPossession.new(:character_id => character.id, :possession_id => "shaper_b").save!
         if(Random.rand(2)==0)
-          return "TODO"
+          return "You craft an angled shaper."
         else
-          return "TODO"
+          return "Your angled shaper breaks most of the way through the process. It's ruined."
         end
       else
-        return "TODO"
+        return "You don't have the materials to craft a shaper."
       end
     },
     :base_cost => 7,
@@ -447,17 +447,17 @@ Action.new("craft_shaper_c",
   { :name=>"Craft Pronged Shaper",
     :description=>"Craft a simple pronged shaping tool to aid in crafting.",
     :result => lambda { |character, character_action|
-      if(character.possesses?("tomatunk"))
+      if(character.possesses?("tomatunk") && character.possesses?("dolait"))
         character.character_possessions.where(:possession_id=>"dolait").first.destroy!
         character.character_possessions.where(:possession_id=>"tomatunk").first.destroy!
         CharacterPossession.new(:character_id => character.id, :possession_id => "shaper_c").save!
         if(Random.rand(2)==0)
-          return "TODO"
+          return "You craft a pronged shaper."
         else
-          return "TODO"
+          return "Your pronged shaper breaks most of the way through the process. It's ruined."
         end
       else
-        return "TODO"
+        return "You don't have the materials to craft a shaper."
       end
     },
     :base_cost => 7,
